requirements.txt
flask
opencv-python
numpy
pillow
pydicom
pandas
matplotlib
simpleitk
scikit-image
tqdm
pyserial
keyboard
pygetwindow
